refactor(package): 패키지 구성
