+++ src/main/java/order/model/fee/DeliveryFeeCalculator.java
@@ -1,4 +1,4 @@
-package order;
+package order.model.fee;
 
 import java.util.Map;
