+++ src/main/java/order/util/InputParser.java
@@ -1,6 +1,5 @@
-package order;
+package order.util;
 
-import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.Map;
