+++ src/test/java/order/model/DeliveryFeeCalculatorTest.java
@@ -1,5 +1,7 @@
-package order;
+package order.model;
 
+import order.model.fee.DeliveryFeeCalculator;
+import order.model.fee.FeeRange;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
