+++ src/test/java/order/util/OrderConverterTest.java
@@ -1,4 +1,4 @@
-package order;
+package order.util;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
