+++ src/test/java/order/view/InputViewTest.java
@@ -1,4 +1,4 @@
-package order;
+package order.view;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
